update pysoda timestamps to include timezone
Given that SPARC has data coming from multiple continents it is
critical that all timestamps include a timezone. This commit updates
pysoda to obtain and format timestamps in the same way that the sparc
curation pipelines do and adds augpathlib as a depdnency since it is
the library that the curation pipelines use for obtaining files system
metadata and formatting timestamps.