organize_dataset.py fix timestamp format and use createdAt
Timestamps sourced directly from the pennsieve packages endpoint are
now formatted to match curation pipeline expectations.

On most filesystems mtime is the only reliable time for files since
file creation time is not tracked. On Pennsieve this is not the case
because files (packages) are immutable, therefore use createdAt to get
the earliest time that a file was known to be in the system.